Make Tanimoto calculation use 64 bit operations.
Depends on -mpopcnt (#186).

Also on top of multi_fastsearch, which is not necessary but is hard
to see the benefit without that.

No attempt was make to make this work on other platforms.

This reduces the time needed to search a 6 million structure fs file 100 times
from 45 seconds to 27 seconds.