Make it work with fingerprint sizes not a multiple fo 64.
Use two loops, one to count by 64bits on machines that support it, and
the second to finish the job where not a multiple of 64 or do all
the work on machines without a popcountll.

Now I get identical output (against master) for searching a file created by:
obabel foo.can -ofs -xfFP3 -xN32

Where as before the popcountll gave the wrong results on these 32 bit
fingerprints.